Unify files.
